Adds CLI Wrapper
Adds `bin` entry to `package.json` pointing to `./bin/backstop.js`. This is
the script that will be linked into `./node_modules/.bin` if installing
locally or linked to `/usr/local/bin` or wherever depending on your
operating system if installing globally.

Adds `./bin/backstop.js` script. It simply parses args and calls a
function.

Adds `./lib/backstop.js` module. This is the wrapper around the gulp
commands. It uses `npm.commands.run` so that later when/if we install
casperjs and phantomjs via npm, they will be in the bin path.

Adds `gulp` script to `package.json` so that in the backstop module, npm
run will be able to run it.